Changed number of bytes to transmit in swift_nap_xfer_blocking from u8 to u16

+++ src/swift_nap_io.h
@@ -92,7 +92,7 @@ typedef struct {
 
 void swift_nap_setup();
 void swift_nap_reset();
-void swift_nap_xfer_blocking(u8 spi_id, u8 n_bytes, u8 data_in[], const u8 data_out[]);
+void swift_nap_xfer_blocking(u8 spi_id, u16 n_bytes, u8 data_in[], const u8 data_out[]);
 
 u8 swift_nap_conf_done();
 u8 swift_nap_hash_rd_done();
